backend/requirements.txt
# --- Framework principal ---
fastapi==0.104.1
uvicorn[standard]==0.24.0

# --- ORM y Base de datos ---
sqlalchemy>=2.0
mysql-connector-python==8.2.0

# --- Utilidades ---
python-dotenv==1.0.0
python-multipart==0.0.6  # para manejar formularios / uploads de archivos

# --- Datos / Archivos ---
pandas>=2.0
openpyxl>=3.1.2

# --- Opcional (para logs, validación, seguridad) ---
pydantic==2.5.0
